fix(resty.dns.client): fix the leak of UDP sockets in resty.dns.client
In `resty.dns.client` module a resolver is created for domain query.
When `resolver:new()` is called, a UDP socket is created. In 2.x, the
UDP socket is released quickly, while in 3.x, it has to take several
seconds to release. If the ttl of dns reponses is shorter than the time
it waits, the leak will happen. In this PR, a forcible release is added.

FTI-4962